+++ EngineQ/EngineQ/Graphics/Mesh.hpp
@@ -48,7 +48,7 @@ namespace EngineQ
 				glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->vbo[2]);
 				glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(GLuint), indices.data(), GL_STATIC_DRAW);
 				glBindBuffer(GL_ARRAY_BUFFER, NULL);
-				indicesCount = indices.size();
+				indicesCount = static_cast<GLuint>(indices.size());
 			}
 
 			Mesh(const Resources::Model::Mesh& mesh);
